Makes it possible to run with only time series as output
This PR makes all entries in the output section of the YAML config file
optional, letting us run configurations with only time series, etc.

This is a pretty simple fix, but we might want to add some more tests at
some point that test whether all these output configurations are doing
what we think they are. Perhaps this is easiest to do as complaints
come in for the moment. :-)

Closes #100