Separates out CEED resource/context management.
This PR moves the operative CEED resource and context out of the dycore and
into a global context, since there's no situation I can think of in which we
would ever support multiple CEED resources. This allows us to gather some logic
and separate concerns.

Also, this PR includes a minor modification to some driver code regarding
datasets (which are still under construction, as I understand it). The ability
to declare variables within case blocks of switch statements is supported by
C23, but many compilers are still using the C17 form of the language, so I've
made the code C17-compliant.